+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- HTML output for analysis
 ### Changed
-- Add a new arity to `make-widget-async` to provide a different widget shape.
 ### Removed
 ### Fixed